Added book read confirmation fixes #23

+++ front/src/helpers/book.ts
@@ -1,31 +1,41 @@
 import {api} from "../api/api";
-import {Book} from "../types/book";
+import {Book, BookWithProgress} from "../types/book";
 import {SerieWithProgress} from "../types/serie";
 
 export async function nextBook(book:Book):Promise<void> {
-    console.log(book);
-    const nextBookId = await api.get<{id:string}>(`books/${book._id}/next`);
+    const foundBook = await api.get<BookWithProgress>(`books/${book._id}/next`);
 
-    if (nextBookId.id === "end") {
+    if (foundBook.status === "completed") {
+        if (!confirm("Yas has leído este volumen. ¿Quieres iniciar un nuevo progreso de lectura?")) return;
+    }
+
+    if (foundBook._id === "end") {
         window.location.href = `/app/series/${book.serie}`;
         return;
     }
-    window.location.href = `/reader/${nextBookId.id}`;
+    window.location.href = `/reader/${foundBook._id}`;
     return;
 }
 
 export async function prevBook(book:Book):Promise<void> {
-    const nextBookId = await api.get<{id:string}>(`books/${book._id}/prev`);
+    const foundBook = await api.get<BookWithProgress>(`books/${book._id}/prev`);
 
-    if (nextBookId.id === "start") {
+    if (foundBook.status === "completed") {
+        if (!confirm("Yas has leído este volumen. ¿Quieres iniciar un nuevo progreso de lectura?")) return;
+    }
+
+    if (foundBook._id === "start") {
         window.location.href = `/app/series/${book.serie}`;
         return;
     }
-    window.location.href = `/reader/${nextBookId.id}`;
+    window.location.href = `/reader/${foundBook._id}`;
     return;
 }
 
 export function iBook(serieData:SerieWithProgress):void {
+    if (serieData.unreadBooks === 0) {
+        if (!confirm("Yas has leído este volumen. ¿Quieres iniciar un nuevo progreso de lectura?")) return;
+    }
     window.localStorage.setItem("origin", window.location.pathname);
     window.location.href = `/reader/${serieData.currentBook}`;
 }
